refactor: 알림 권한 요청 퀴즈 화면에서 로그인 화면으로 이전
